Adds Anthony's feedback
We changed line L36 because we think is redundant to L32.
Removed 'Manage all the things' in the solutions requirments as we could
not find the value to it.
Mentioned the lack of official Ops Manager API.

Overall Anthony likes the talk. He made a good suggestion on a key
aspect that we have not mentined enough. Related to changes. As changes
can break stuff we have two options, embraceing them or not. If we dont
we have fewer changes with higher risk, if we do we have many changes
with lower risk and we iterate to get better at it. Thats his take